[DEV] counting characters, undo/redo
