--- src/util/caseHandler.ts
@@ -1,54 +1,0 @@
-import {Case} from 'constant/Cases';
-
-/**
- * Change the case with more option
- * @param  {string} action 	constant from Cases
- * @param  {string} text 	text to change
- * @return {string} 		changedText changed text
- */
-export const changeComplexCase = (action:string, text:string):string => {
-	let changedText:string = '',
-		caseRegex:RegExp = /./g,
-		lowerRegex = new RegExp('\\p{Lower}', 'u');
-
-	// camel case and capitalize
-	if (action === Case.camel || action === Case.capital) {
-		changedText = text.toLowerCase();
-		caseRegex = (action === Case.camel) ? /\s\p{Letter}/gu : /(^|\s)\p{Letter}/gu;
-	}
-
-	// case inversion
-	if (action === Case.inversion) {
-		changedText = text;
-		caseRegex = /\p{Letter}/gu;
-	}
-
-	// change text
-	if (changedText) 
-	{
-		changedText = changedText.replaceAll(caseRegex, (letter:string):string => (lowerRegex.test(letter)) ? letter.toUpperCase() : letter.toLowerCase());
-		// delete whitespaces for camel case
-		if (action === Case.camel) changedText = changedText.replaceAll(/\s/g, '');
-	}
-	return changedText;
-}
-
-/**
- * [description]
- * @param  {[type]} string [description]
- * @return {[type]}        [description]
- */
-export const changeCase = (action:string, text:string):string => {
-	switch (action) {
-		case Case.upper:
-			text = text.toUpperCase();
-			break;
-		case Case.lower:
-			text = text.toLowerCase();
-			break;
-		default:
-			text = changeComplexCase(action, text);
-	}
-
-	return text;
-}
